Cleanup actions and workflows

+++ packs/st2cd/actions/teardown_mistral_itests.sh
@@ -26,14 +26,9 @@ fi
 service mistral stop
 # Kill mistral
 MISTRAL_PID=`ps auxww | grep mistral | grep -v grep | awk '{print $2}'`
-kill -s TERM $MISTRAL_PID
-
-if [[ $? == 0 ]]
+if [ ! -z "$MISTRAL_PID" ]
 then
-    echo "SUCCESS: Killed mistral process with PID ${MISTRAL_PID}"
-else
-    echo "ERROR: Could not kill mistral process with PID ${MISTRAL_PID}"
-    exit 1
+    kill -s TERM $MISTRAL_PID
 fi
 
 # Wipe clean dbs
